+++ src/Terrabuild/Contracts/Api.fs
@@ -28,9 +28,9 @@ type CommitGraph = {
 
 type IApiClient =
     abstract StartBuild: Unit -> Unit
-    abstract UploadBuildGraph: graphHash:string -> nodes:BuildGraphNode list -> Unit
+    abstract UploadBuildGraph: graphHash:string -> environment:string -> nodes:BuildGraphNode list -> Unit
     abstract CompleteBuild: success:bool -> Unit
     abstract AddArtifact: project:string -> projectName:string option -> target:string -> projectHash:string -> targetHash:string  -> files:string list -> success:bool -> startedAt:DateTime -> endedAt:DateTime -> Unit
     abstract UseArtifact: projectHash:string -> hash:string -> Unit
     abstract GetArtifact: path:string -> Uri
-    abstract GetCommitGraph: repository:string -> commit:string -> CommitGraph
+    abstract GetCommitGraph: repository:string -> commit:string -> environment:string -> CommitGraph